Consolidate TestCases and Hardcode baseUrl
 - There is no need to maintain two separate testcases,
   so combined them following discussion with crynobone.

 - Fix the baseUrl used for running the tests. Almost
   everyone will use the default setting, so this will
   reduce the change of incorrect setup.

   It remains possible to customise the host/port if
   required.
