--- src/Foundation.php
@@ -1,71 +1,0 @@
-<?php
-
-namespace Orchestra\Testbench\Dusk;
-
-use Closure;
-use Exception;
-use Orchestra\Testbench\TestCase;
-use Facebook\WebDriver\Remote\RemoteWebDriver;
-use Facebook\WebDriver\Remote\DesiredCapabilities;
-
-abstract class Foundation extends TestCase
-{
-    use Concerns\CanServeSite,
-        Concerns\ProvidesBrowser;
-
-    /**
-     * Register the base URL with Dusk.
-     *
-     * @return void
-     */
-    protected function setUp()
-    {
-        parent::setUp();
-
-        $this->setUpTheBrowserEnvironment();
-    }
-
-    /**
-     * Get base path.
-     *
-     * @return string
-     */
-    protected function getBasePath()
-    {
-        return __DIR__.'/../laravel';
-    }
-
-    /**
-     * Create the RemoteWebDriver instance.
-     *
-     * @return \Facebook\WebDriver\Remote\RemoteWebDriver
-     */
-    protected function driver()
-    {
-        return RemoteWebDriver::create(
-            'http://localhost:9515', DesiredCapabilities::chrome()
-        );
-    }
-
-    /**
-     * Determine the application's base URL.
-     *
-     * @var string
-     */
-    protected function baseUrl()
-    {
-        return config('app.url');
-    }
-
-    /**
-     * Get a callback that returns the default user to authenticate.
-     *
-     * @throws \Exception
-     *
-     * @return \Closure
-     */
-    protected function user()
-    {
-        throw new Exception('User resolver has not been set.');
-    }
-}
