+++ tests/Unit/ResolvesProjectRootTest.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace Orchestra\Testbench\Dusk\Tests\Unit;
+
+use PHPUnit\Framework\TestCase;
+use Orchestra\Testbench\Dusk\TestCase as TestbenchDuskTestCase;
+
+class TestCaseTest extends TestCase
+{
+    /** @test */
+    public function it_finds_the_correct_path_for_the_browser_tests()
+    {
+        $this->assertEquals(
+            '/home/person/code/testbench-dusk/tests/Browser',
+            (new DummyTestCase())->resolveBrowserTestsPath('/home/person/code/testbench-dusk/src')
+        );
+
+        $this->assertEquals(
+            '/home/person/code/project/tests/Browser',
+            (new DummyTestCase())->resolveBrowserTestsPath('/home/person/code/project/vendor/orchestra/testbench-dusk/src')
+        );
+    }
+}
+
+class DummyTestCase extends TestbenchDuskTestCase
+{
+    public function resolveBrowserTestsPath($path = __DIR__)
+    {
+        return parent::resolveBrowserTestsPath($path);
+    }
+
+    // Added so PHPUnit doesn't warn about not tests being found in this class.
+    public function testQuietly()
+    {
+        $this->assertTrue(true);
+    }
+}
